move Specification classes to separate file
