+++ tests/test_api.py
@@ -7,8 +7,8 @@ from yaml import YAMLError
 
 import pytest
 from connexion import FlaskApi
-from connexion.apis.abstract import canonical_base_path
 from connexion.exceptions import InvalidSpecification, ResolverError
+from connexion.spec import canonical_base_path
 from mock import MagicMock
 
 TEST_FOLDER = pathlib.Path(__file__).parent
